refactor: View를 만드는 메서드를 별도의 타입으로 분리, 코딩 컨벤션 정리
