+++ Calculator/Calculator/AppDelegate.swift
@@ -1,9 +1,3 @@
-//
-//  Calculator - AppDelegate.swift
-//  Created by yagom. 
-//  Copyright © yagom. All rights reserved.
-// 
-
 import UIKit
 
 @main
