+++ Calculator/Calculator/SceneDelegate.swift
@@ -1,9 +1,3 @@
-//
-//  Calculator - SceneDelegate.swift
-//  Created by yagom. 
-//  Copyright © yagom. All rights reserved.
-// 
-
 import UIKit
 
 class SceneDelegate: UIResponder, UIWindowSceneDelegate {
