+++ Calculator/CalculatorTests/ExpressionParserTDD.swift
@@ -1,10 +1,3 @@
-//
-//  ExpressionParserTDD.swift
-//  CalculatorTests
-//
-//  Created by LeeChiheon on 2022/05/20.
-//
-
 import XCTest
 
 class ExpressionParserTDD: XCTestCase {
