+++ Calculator/CalculatorTests/FormulaTDD.swift
@@ -1,10 +1,3 @@
-//
-//  FormulaTDD.swift
-//  CalculatorTests
-//
-//  Created by LeeChiheon on 2022/05/20.
-//
-
 import XCTest
 
 class FormulaTDD: XCTestCase {
