+++ Calculator/CalculatorTests/StringExtensionTDD.swift
@@ -1,10 +1,3 @@
-//
-//  StringExtensionTDD.swift
-//  CalculatorTests
-//
-//  Created by LeeChiheon on 2022/05/20.
-//
-
 import XCTest
 
 class StringExtensionTDD: XCTestCase {
